requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
peft>=0.4.0
pillow>=9.0.0
torchvision>=0.15.0
requests>=2.28.0

# For mixed precision training
nvidia-ml-py3>=7.352.0

# Optional for faster tokenization
sentencepiece>=0.1.97

# Optional: For logging and monitoring
tensorboard>=2.10.0
